Address PR feedback; remove limit on aiSummary feature flag
