+++ packages/server/graphql/mutations/helpers/generateWholeMeetingSentimentScore.ts
@@ -2,14 +2,9 @@ import {DataLoaderWorker} from '../../graphql'
 
 const generateWholeMeetingSentimentScore = async (
   meetingId: string,
-  facilitatorUserId: string,
   dataLoader: DataLoaderWorker
 ) => {
-  const [facilitator, reflections] = await Promise.all([
-    dataLoader.get('users').loadNonNull(facilitatorUserId),
-    dataLoader.get('retroReflectionsByMeetingId').load(meetingId)
-  ])
-  if (!facilitator.featureFlags.includes('aiSummary')) return undefined
+  const reflections = await dataLoader.get('retroReflectionsByMeetingId').load(meetingId)
   return (
     reflections.reduce((_prev, reflection) => _prev + reflection.sentimentScore, 0.0) /
     reflections.length
