+++ packages/server/graphql/mutations/helpers/getReflectionEntities.ts
@@ -1,28 +1,7 @@
 import getGoogleLanguageManager from '../../../getGoogleLanguageManager'
-import {GoogleErrorResponse} from '../../../GoogleLanguageManager'
-import sendToSentry from '../../../utils/sendToSentry'
+import manageGoogleNLPErrorResponse from './manageGoogleNLPErrorResponse'
 import addLemmaToEntities from './autoGroup/addLemmaToEntities'
 import sanitizeAnalyzedEntitiesResponse from './autoGroup/sanitizeAnalyzedEntititesResponse'
-
-const manageErrorResponse = <T>(res: T) => {
-  if (Array.isArray(res)) {
-    const [firstError] = res
-    if (firstError) {
-      const {error} = firstError
-      if (error) {
-        const {message} = error
-        if (message !== 'No JWT provided') {
-          const re = /language \S+ is not supported/
-          if (!re.test(message)) {
-            sendToSentry(new Error(`Grouping Error: Google NLP: ${message}`))
-          }
-        }
-      }
-    }
-    return null
-  }
-  return res as T extends GoogleErrorResponse ? never : T
-}
 
 const getReflectionEntities = async (plaintextContent: string) => {
   if (!plaintextContent) return []
@@ -31,8 +10,8 @@ const getReflectionEntities = async (plaintextContent: string) => {
     manager.analyzeEntities(plaintextContent),
     manager.analyzeSyntax(plaintextContent)
   ])
-  const reflectionResponse = manageErrorResponse(res[0])
-  const reflectionSyntax = manageErrorResponse(res[1])
+  const reflectionResponse = manageGoogleNLPErrorResponse(res[0])
+  const reflectionSyntax = manageGoogleNLPErrorResponse(res[1])
   // for each entity, look in the tokens array to first the first word of the entity
   // for each word in the entity, make sure that the next token points to it, else continue, return entity starting index
   // take the lemma of the last word and recompute the entity based on that lemma
@@ -43,12 +22,4 @@ const getReflectionEntities = async (plaintextContent: string) => {
   return addLemmaToEntities(sanitizedReflectionResponse, reflectionSyntax)
 }
 
-const getReflectionSentimentScore = async (plaintextContent: string) => {
-  if (!plaintextContent) return 0.0
-  const manager = getGoogleLanguageManager()
-  const res = await manager.analyzeSentiment(plaintextContent)
-  const reflectionSentiment = manageErrorResponse(res)
-  return reflectionSentiment?.documentSentiment.score ?? 0.0
-}
-
-export {getReflectionEntities, getReflectionSentimentScore}
+export default getReflectionEntities
